+++ visao/grafico.h
@@ -9,8 +9,8 @@
 #define DISTANCIA 1.0
 #define QTD_DIST_UM 0.1
 
-void planoCarteziano (float &esquerda, float &direita, double baixo, double cima);
-void mostrarFuncao (expre::expre *expressao, float inter[2]);
+void planoCarteziano (double &esquerda, double &direita, double baixo, double cima);
+void mostrarFuncao (expre::expre *expressao, double inter[2]);
 void mostrarArea (std::vector <interp> &interpol);
 void mostrarInterpolacao (std::vector <interp> &interpol, std::vector <interp> &pontosInterpol);
